+++ examples/counter/src/main.rs
@@ -20,7 +20,6 @@ fn main() {
     let mut node = Node::new("text".to_string());
     node.add_component(text);
     node.move_to(Vector2 { x: 0.0, y: 0.0 });
-    node.set_is_static(true);
     node.set_id(2);
 
     scene.add_node(node);
@@ -63,7 +62,6 @@ fn main() {
         let mut node = Node::new("text".to_string());
         node.add_component(text);
         node.move_to(pos);
-        node.set_is_static(true);
         node.set_id(2);
 
         game.get_scene_mut().unwrap().add_node(node);